Optimization: Optimize code #74

+++ app/src/main/java/me/itangqi/buildingblocks/widget/ScrollingFABBehavior.java
@@ -29,7 +29,7 @@ public class ScrollingFABBehavior extends FloatingActionButton.Behavior {
                 CoordinatorLayout.LayoutParams lp = (CoordinatorLayout.LayoutParams) fab.getLayoutParams();
                 int fabBottomMargin = lp.bottomMargin;
                 int distanceToScroll = fab.getHeight() + fabBottomMargin;
-                float ratio = (float)dependency.getY()/(float)toolbarHeight;
+                float ratio = dependency.getY()/(float)toolbarHeight;
                 fab.setTranslationY(-distanceToScroll * ratio);
         }
         return returnValue;
